requirements.txt
agno
google-genai
python-dotenv
youtube-transcript-api